Rename projection files
