+++ src/render/draw_symbol.test.ts
@@ -15,7 +15,7 @@ import {type IReadonlyTransform} from '../geo/transform_interface';
 import type {EvaluationParameters} from '../style/evaluation_parameters';
 import type {SymbolLayerSpecification} from '@maplibre/maplibre-gl-style-spec';
 import {type Style} from '../style/style';
-import {MercatorProjection} from '../geo/projection/mercator';
+import {MercatorProjection} from '../geo/projection/mercator_projection';
 import type {ProjectionData} from '../geo/projection/projection_data';
 
 vi.mock('./painter');
