+++ test/bench/benchmarks/covering_tiles_globe.ts
@@ -1,6 +1,6 @@
 import Benchmark from '../lib/benchmark';
 import { GlobeTransform } from '../../../src/geo/projection/globe_transform';
-import { GlobeProjection } from '../../../src/geo/projection/globe';
+import { GlobeProjection } from '../../../src/geo/projection/globe_projection';
 import { LngLat } from '../styles';
 import { coveringTiles } from '../../../src/geo/projection/covering_tiles';
