+++ test/bench/benchmarks/symbol_collision_box.ts
@@ -8,7 +8,7 @@ import {SingleCollisionBox} from '../../../src/data/bucket/symbol_bucket';
 import {EXTENT} from '../../../src/data/extent';
 import {MercatorTransform} from '../../../src/geo/projection/mercator_transform';
 import {mat4} from 'gl-matrix';
-import {GlobeProjection} from '../../../src/geo/projection/globe';
+import {GlobeProjection} from '../../../src/geo/projection/globe_projection';
 import {GlobeTransform} from '../../../src/geo/projection/globe_transform';
 
 type TestSymbol = {
